AI ok, no loop case found need to continue check

+++ include/lem_ipc.h
@@ -167,7 +167,7 @@ void		team_handling(uint32_t *array, uint32_t team_id, int8_t add);
 int8_t		remove_msg_queue(t_ipc *ipc);
 uint32_t 	extract_msg(t_ipc *ipc, t_player *player);
 int8_t		send_msg(t_ipc *ipc, t_player *player, uint32_t data);
-int8_t		clear_msg_queue(t_ipc *ipc, uint32_t team_id);
+int8_t clear_msg_queue(t_ipc *ipc, long team_id);
 /* init semaphore */
 int			init_game(t_ipc *ipc, char *path, int8_t allow);
